fix(test): 누락 테스트 추가
