+++ src/main/java/model/result/ParticipantsResult.java
@@ -7,6 +7,7 @@ import java.util.Map;
 import java.util.Objects;
 
 public class ParticipantsResult {
+    protected static final String NOT_EXIST_PARTICIPANT = "존재하지 않는 참가자 입니다.";
     private final Map<Participant, Result> participantsResult;
 
     public ParticipantsResult(Map<Participant, Result> participantsResult) {
@@ -17,8 +18,11 @@ public class ParticipantsResult {
         return new LinkedHashMap<>(participantsResult);
     }
 
-    public Result getResult(Participant participant){
-        return participantsResult.get(participant);
+    public Result getResult(Participant participant) {
+        if (participantsResult.containsKey(participant)) {
+            return participantsResult.get(participant);
+        }
+        throw new IllegalArgumentException(NOT_EXIST_PARTICIPANT);
     }
 
     @Override
