+++ src/main/java/model/result/Results.java
@@ -2,7 +2,6 @@ package model.result;
 
 import model.position.Position;
 
-import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
@@ -29,7 +28,4 @@ public class Results {
         return this.results.get(position);
     }
 
-    public List<Result> getResults() {
-        return new ArrayList<>(results.values());
-    }
 }
